third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-end-after-get-set.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<style>
.gridItemWithPositiveInteger {
    -webkit-grid-end: 10;
    -webkit-grid-after: 15;
}
.gridItemWithNegativeInteger {
    -webkit-grid-end: -10;
    -webkit-grid-after: -15;
}
.gridItemWithBeforeSpan {
    -webkit-grid-end: span 2;
    -webkit-grid-after: span 9;
}
.gridItemWithAfterSpan {
    -webkit-grid-end: 2 span;
    -webkit-grid-after: 9 span;
}
.gridItemWithOnlySpan {
    -webkit-grid-end: span;
    -webkit-grid-after: span;
}
.gridItemWithAuto {
    -webkit-grid-end: auto;
    -webkit-grid-after: auto;
}
.gridItemWithString {
    -webkit-grid-end: "first";
    -webkit-grid-after: "last";
}
.gridItemWithSpanString {
    -webkit-grid-end: "first" span;
    -webkit-grid-after: span "last";
}
.gridItemWithSpanNumberString {
    -webkit-grid-end: 2 "first" span;
    -webkit-grid-after: "last" 3 span;
}
</style>
<script src="resources/grid-item-column-row-parsing-utils.js"></script>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<!-- The first has no properties set on it. -->
<div id="gridElement"></div>
<div class="gridItemWithPositiveInteger" id="gridItemWithPositiveInteger"></div>
<div class="gridItemWithNegativeInteger" id="gridItemWithNegativeInteger"></div>
<div class="gridItemWithBeforeSpan" id="gridItemWithBeforeSpan"></div>
<div class="gridItemWithAfterSpan" id="gridItemWithAfterSpan"></div>
<div class="gridItemWithOnlySpan" id="gridItemWithOnlySpan"></div>
<div class="gridItemWithAuto" id="gridItemWithAutoElement"></div>
<div class="gridItemWithString" id="gridItemWithStringElement"></div>
<div class="gridItemWithSpanString" id="gridItemWithSpanStringElement"></div>
<div class="gridItemWithSpanNumberString" id="gridItemWithSpanNumberStringElement"></div>
<script>
    description('Test that setting and getting grid-end and grid-after works as expected');

    debug("Test getting -webkit-grid-end and -webkit-grid-after set through CSS");
    testColumnRowCSSParsing("gridElement", "auto / auto", "auto / auto");
    testColumnRowCSSParsing("gridItemWithPositiveInteger", "auto / 10", "auto / 15");
    testColumnRowCSSParsing("gridItemWithNegativeInteger", "auto / -10", "auto / -15");
    testColumnRowCSSParsing("gridItemWithBeforeSpan", "auto / span 2", "auto / span 9");
    testColumnRowCSSParsing("gridItemWithAfterSpan", "auto / span 2", "auto / span 9");
    testColumnRowCSSParsing("gridItemWithOnlySpan", "auto / span 1", "auto / span 1");
    testColumnRowCSSParsing("gridItemWithAutoElement", "auto / auto", "auto / auto");
    testColumnRowCSSParsing("gridItemWithStringElement", "auto / 1 first", "auto / 1 last");
    testColumnRowCSSParsing("gridItemWithSpanStringElement", "auto / span 1 first", "auto / span 1 last");
    testColumnRowCSSParsing("gridItemWithSpanNumberStringElement", "auto / span 2 first", "auto / span 3 last");

    debug("");
    debug("Test the initial value");
    var element = document.createElement("div");
    document.body.appendChild(element);
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");

    debug("");
    debug("Test getting and setting grid-end and grid-after through JS");
    testEndAfterJSParsing("18", "66");
    testEndAfterJSParsing("-55", "-40");
    testEndAfterJSParsing("'first'", "'last'", "1 first", "1 last");
    testEndAfterJSParsing("span 7", "span 2");
    testEndAfterJSParsing("span 'first'", "span 'last'", "span 1 first", "span 1 last");
    testEndAfterJSParsing("auto", "auto");

    debug("");
    debug("Test setting grid-start and grid-before to 'inherit' through JS");
    testEndAfterInheritJSParsing("inherit", "18");
    testEndAfterInheritJSParsing("2", "inherit");
    testEndAfterInheritJSParsing("inherit", "inherit");

    debug("");
    debug("Test setting grid-start and grid-before to 'initial' through JS");
    testEndAfterInitialJSParsing();

    debug("");
    debug("Test setting grid-end and grid-after back to 'auto' through JS");
    element.style.webkitGridEnd = "18";
    element.style.webkitGridAfter = "66";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / 18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'66'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / 66'");
    element.style.webkitGridEnd = "auto";
    element.style.webkitGridAfter = "auto";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
